updated to latest bzr branch

+++ container/lxc/lxc.go
@@ -10,9 +10,10 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/juju/loggo"
 	"launchpad.net/golxc"
-	"launchpad.net/loggo"
 
+	"launchpad.net/juju-core/agent"
 	"launchpad.net/juju-core/container"
 	"launchpad.net/juju-core/environs/cloudinit"
 	"launchpad.net/juju-core/instance"
@@ -51,12 +52,22 @@ var _ container.Manager = (*containerManager)(nil)
 // NewContainerManager returns a manager object that can start and stop lxc
 // containers. The containers that are created are namespaced by the name
 // parameter.
-func NewContainerManager(conf container.ManagerConfig) container.Manager {
-	logdir := "/var/log/juju"
-	if conf.LogDir != "" {
-		logdir = conf.LogDir
+func NewContainerManager(conf container.ManagerConfig) (container.Manager, error) {
+	name := conf[container.ConfigName]
+	delete(conf, container.ConfigName)
+	if name == "" {
+		return nil, fmt.Errorf("name is required")
 	}
-	return &containerManager{name: conf.Name, logdir: logdir}
+	logDir := conf[container.ConfigLogDir]
+	delete(conf, container.ConfigLogDir)
+	if logDir == "" {
+		logDir = agent.DefaultLogDir
+	}
+	for k, v := range conf {
+		logger.Warningf(`Found unused config option with key: "%v" and value: "%v"`, k, v)
+	}
+
+	return &containerManager{name: name, logdir: logDir}, nil
 }
 
 func (manager *containerManager) StartContainer(
@@ -109,12 +120,17 @@ func (manager *containerManager) StartContainer(
 		return nil, nil, err
 	}
 	logger.Tracef("lxc container created")
-	// Now symlink the config file into the restart directory.
-	containerConfigFile := filepath.Join(LxcContainerDir, name, "config")
-	if err := os.Symlink(containerConfigFile, restartSymlink(name)); err != nil {
-		return nil, nil, err
+	// Now symlink the config file into the restart directory, if it exists.
+	// This is for backwards compatiblity. From Trusty onwards, the auto start
+	// option should be set in the LXC config file, this is done in the networkConfigTemplate
+	// function below.
+	if useRestartDir() {
+		containerConfigFile := filepath.Join(LxcContainerDir, name, "config")
+		if err := os.Symlink(containerConfigFile, restartSymlink(name)); err != nil {
+			return nil, nil, err
+		}
+		logger.Tracef("auto-restart link created")
 	}
-	logger.Tracef("auto-restart link created")
 
 	// Start the lxc container with the appropriate settings for grabbing the
 	// console output and a log file.
@@ -140,10 +156,12 @@ func (manager *containerManager) StartContainer(
 func (manager *containerManager) StopContainer(instance instance.Instance) error {
 	name := string(instance.Id())
 	lxcContainer := LxcObjectFactory.New(name)
-	// Remove the autostart link.
-	if err := os.Remove(restartSymlink(name)); err != nil {
-		logger.Errorf("failed to remove restart symlink: %v", err)
-		return err
+	if useRestartDir() {
+		// Remove the autostart link.
+		if err := os.Remove(restartSymlink(name)); err != nil {
+			logger.Errorf("failed to remove restart symlink: %v", err)
+			return err
+		}
 	}
 	if err := lxcContainer.Destroy(); err != nil {
 		logger.Errorf("failed to destroy lxc container: %v", err)
@@ -198,7 +216,12 @@ lxc.network.flags = up
 `
 
 func networkConfigTemplate(networkType, networkLink string) string {
-	return fmt.Sprintf(networkTemplate, networkType, networkLink)
+	networkConfig := fmt.Sprintf(networkTemplate, networkType, networkLink)
+	if !useRestartDir() {
+		networkConfig += "lxc.start.auto = 1\n"
+		logger.Tracef("Setting auto start to true in lxc config.")
+	}
+	return networkConfig
 }
 
 func generateNetworkConfig(network *container.NetworkConfig) string {
@@ -225,4 +248,19 @@ func writeLxcConfig(network *container.NetworkConfig, directory, logdir string) 
 		return "", err
 	}
 	return configFilename, nil
+}
+
+// useRestartDir is used to determine whether or not to use a symlink to the
+// container config as the restart mechanism.  Older versions of LXC had the
+// /etc/lxc/auto directory that would indicate that a container shoud auto-
+// restart when the machine boots by having a symlink to the lxc.conf file.
+// Newer versions don't do this, but instead have a config value inside the
+// lxc.conf file.
+func useRestartDir() bool {
+	if _, err := os.Stat(LxcRestartDir); err != nil {
+		if os.IsNotExist(err) {
+			return false
+		}
+	}
+	return true
 }
